src/api/requirements.txt
fastapi
uvicorn
azure-search-documents
azure-identity
azure-core
requests
python-dotenv